2009-11-26  Sebastien Pouliot  <sebastien@ximian.com>
	* Random.cs: Fix case where the diff between min and max
	is very large in Next(int,int). Fix bug #558593
	[Backport r146993]


svn path=/branches/mono-2-4/mcs/; revision=146995

+++ mcs/class/corlib/System/Random.cs
@@ -124,7 +124,7 @@ namespace System
 			if (diff <= 1)
 				return minValue;
 
-			return (int)(Sample () * diff) + minValue;
+			return (int)((uint)(Sample () * diff) + minValue);
 		}
 
 		public virtual void NextBytes (byte [] buffer)
